+++ src/main/java/lotto/domain/lotto/LottoNumbers.java
@@ -1,7 +1,7 @@
 package lotto.domain.lotto;
 
+import lotto.constants.ErrorConsts;
 import lotto.constants.LottoConsts;
-import lotto.exception.constants.ErrorConsts;
 
 import java.util.HashSet;
 import java.util.List;
